Extract ACR methods to a separate class
